requirements.txt
Django==5.2.5
google-generativeai==0.8.5
python-decouple==3.8
requests==2.32.4

